updated old commented code for the Manta entity, and also addded entity definition for the Creeper (not the explodey kind, the kind that is effectively a long ranged spider)

+++ src/main/java/com/Ultra_Nerd/CodeLyokoLegacy/Blocks/Fluids/LiquidHeliumFluid.java
@@ -47,7 +47,7 @@ public abstract class LiquidHeliumFluid extends LyokoFluid {
 
         @Override
         public int getLevel(final FluidState state) {
-            return state.<Integer>get(LEVEL).intValue();
+            return state.<Integer>get(LEVEL);
         }
 
         @Override
