+++ src/main/java/com/Ultra_Nerd/CodeLyokoLegacy/Blocks/Fluids/UraniumWasteFluid.java
@@ -45,7 +45,7 @@ public abstract class UraniumWasteFluid extends LyokoFluid {
 
         @Override
         public int getLevel(final FluidState state) {
-            return state.<Integer>get(LEVEL).intValue();
+            return state.<Integer>get(LEVEL);
         }
 
         @Override
